#100 replace links from MHeader into parent component
