+++ src/components/molecules/MHeader.vue
@@ -2,44 +2,21 @@
 import { defineComponent } from 'vue';
 import { ATextLogo } from '@/components/atoms';
 import MHeaderMenu from '@/components/molecules/MHeaderMenu.vue';
-import { URL_TO_HOME, URL_TO_FAQ, URL_TO_ABOUT, URL_TO_DONATE, URL_TO_MOBILE } from '@/constants';
+import { URL_TO_HOME } from '@/constants';
 
+import type { PropType } from 'vue';
 import type { Link } from '@/models';
 
 defineComponent({
   name: 'MHeader',
 });
 
-/**
- * Ссылки.
- */
-const links: Link[] = [
-  {
-    text: 'Gamepad',
-    to: URL_TO_HOME,
-    target: '_self',
+defineProps({
+  links: {
+    type: Array as PropType<Link[]>,
+    default: () => [],
   },
-  {
-    text: 'Mobile',
-    to: URL_TO_MOBILE,
-    target: '_self',
-  },
-  {
-    text: 'FAQ',
-    to: URL_TO_FAQ,
-    target: '_self',
-  },
-  {
-    text: 'About',
-    to: URL_TO_ABOUT,
-    target: '_self',
-  },
-  {
-    text: 'Donate',
-    to: URL_TO_DONATE,
-    target: '_self',
-  },
-];
+});
 </script>
 
 <template>
